[수정] 소스코드 패키지로 모듈화
