+++ KuringApp/KuringApp/ContentView.swift
@@ -7,7 +7,9 @@
 
 import SwiftUI
 import NoticeUI
+import SettingsUI
 import NoticeFeatures
+import SettingsFeatures
 import ComposableArchitecture
 
 struct ContentView: View {
