+++ KuringPackage/Sources/Caches/AppIcons.swift
@@ -1,15 +1,26 @@
-//
-//  AppIcons.swift
-//  KuringApp
-//
-//  Created by Jaesung Lee on 2023/10/06.
-//
-
 import UIKit
 import ComposableArchitecture
 
+public enum KuringIcon: String, CaseIterable, Identifiable, Equatable {
+    public var id: String { self.rawValue }
+    case kuring_app
+    case kuring_app_classic
+    case kuring_app_blueprint
+    case kuring_app_sketch
+    
+    public var korValue: String {
+        switch self {
+        case .kuring_app: return "쿠링 기본"
+        case .kuring_app_classic: return "쿠링 클래식"
+        case .kuring_app_blueprint: return "쿠링 블루프린트"
+        case .kuring_app_sketch: return "쿠링 스케치"
+        }
+    }
+}
+
+
 public struct AppIcons: DependencyKey {
-    public static let liveValue: AppIcons  = AppIcons(
+    public static let liveValue: AppIcons = AppIcons(
         changeTo: { icon in
             guard UIApplication.shared.supportsAlternateIcons else { return }
             do {
@@ -21,7 +32,7 @@ public struct AppIcons: DependencyKey {
     )
     
     /// 현재 앱 아이콘
-    var currentAppIcon: KuringIcon {
+    public var currentAppIcon: KuringIcon {
         guard let iconName = UIApplication.shared.alternateIconName else {
             return .kuring_app
         }
@@ -30,7 +41,7 @@ public struct AppIcons: DependencyKey {
     
     /// 앱 아이콘을 변경합니다.
     /// - Important: 메인에서 돌아감.
-    var changeTo: @MainActor (KuringIcon) async -> Void
+    public var changeTo: @MainActor (KuringIcon) async -> Void
 }
 
 extension DependencyValues {
