+++ KuringPackage/Sources/Features/SettingsFeatures/SettingsApp.Destination.swift
@@ -1,31 +1,23 @@
-//
-//  SettingsApp.Destination.swift
-//  KuringApp
-//
-//  Created by 이재성 on 12/29/23.
-//
-
-import SwiftUI
 import SubscriptionFeatures
 import ComposableArchitecture
 
 extension SettingsAppFeature {
     @Reducer
-    struct Destination {
+    public struct Destination {
         @ObservableState
-        enum State: Equatable {
+        public enum State: Equatable {
             case feedback(FeedbackFeature.State)
             case subscription(SubscriptionAppFeature.State)
-            case informationWeb(InformationWebViewFeature.State)
+            case informationWeb(InformationWebFeature.State)
         }
         
-        enum Action: Equatable {
+        public enum Action: Equatable {
             case feedback(FeedbackFeature.Action)
             case subscription(SubscriptionAppFeature.Action)
-            case informationWeb(InformationWebViewFeature.Action)
+            case informationWeb(InformationWebFeature.Action)
         }
         
-        var body: some ReducerOf<Self> {
+        public var body: some ReducerOf<Self> {
             Scope(state: \.feedback, action: \.feedback) {
                 FeedbackFeature()
             }
@@ -33,8 +25,10 @@ extension SettingsAppFeature {
                 SubscriptionAppFeature()
             }
             Scope(state: \.informationWeb, action: \.informationWeb) {
-                InformationWebViewFeature()
+                InformationWebFeature()
             }
         }
+        
+        public init() { }
     }
 }
