+++ KuringPackage/Sources/Features/SettingsFeatures/SettingsApp.Path.swift
@@ -1,29 +1,20 @@
-//
-//  SetttingsApp.Path.swift
-//  KuringApp
-//
-//  Created by Jaesung Lee on 2023/10/06.
-//
-
-import SwiftUI
-import SubscriptionFeatures
 import ComposableArchitecture
 
 extension SettingsAppFeature {
     @Reducer
-    struct Path {
+    public struct Path {
         @ObservableState
-        enum State: Equatable {
+        public enum State: Equatable {
             case openSourceList(OpenSourceListFeature.State)
             case appIconSelector(AppIconSelectorFeature.State)
         }
         
-        enum Action: Equatable {
+        public enum Action: Equatable {
             case openSourceList(OpenSourceListFeature.Action)
             case appIconSelector(AppIconSelectorFeature.Action)
         }
         
-        var body: some ReducerOf<Self> {
+        public var body: some ReducerOf<Self> {
             Scope(state: \.openSourceList, action: \.openSourceList) {
                 OpenSourceListFeature()
             }
@@ -31,5 +22,7 @@ extension SettingsAppFeature {
                 AppIconSelectorFeature()
             }
         }
+        
+        public init() { }
     }
 }
