+++ KuringPackage/Sources/UIKit/SettingsUI/Bundle.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public extension Bundle {
+    static var settings: Bundle { .module }
+}
