+++ KuringPackage/Sources/UIKit/SettingsUI/SettingList.swift
@@ -1,74 +1,12 @@
-//
-//  SettingList.swift
-//  KuringApp
-//
-//  Created by 이재성 on 12/29/23.
-//
-
-import Foundation
+import Caches
+import SwiftUI
+import SettingsFeatures
 import ComposableArchitecture
 
-enum URLLink: String {
-    case team = "https://bit.ly/3v2c5eg"
-    case instagram = "https://instagram.com/kuring.konkuk"
-    case terms = "https://kuring.notion.site/e88095d4d67d4c4c92983fd85cb693b9"
-    case privacy = "https://kuring.notion.site/65ba27f2367044e0be7061e885e7415c"
-    case whatsNew = "https://kuring.notion.site/iOS-eef51c986b7f4320b97424df3f4a5e3c"
-}
-
-@Reducer
-struct SettingListFeature {
-    @ObservableState
-    struct State: Equatable {
-        // TODO: 나중에 디펜던시로
-        var currentAppIcon: KuringIcon?
-        var isCustomAlarmOn: Bool = false
-        
-        init(
-            isCustomAlarmOn: Bool = true,
-            appIcon: KuringIcon? = nil
-        ) {
-            self.isCustomAlarmOn = isCustomAlarmOn
-            
-            @Dependency(\.appIcons) var appIcons
-            self.currentAppIcon = appIcon ?? appIcons.currentAppIcon
-        }
-    }
+public struct SettingList: View {
+    @Bindable public var store: StoreOf<SettingListFeature>
     
-    enum Action: Equatable, BindableAction {
-        case binding(BindingAction<State>)
-        case delegate(Delegate)
-        
-        enum Delegate: Equatable {
-            case showSubscription
-            case showWhatsNew
-            case showTeam
-            case showPrivacyPolicy
-            case showTermsOfService
-            case showInstagram
-            case showFeedback
-        }
-    }
-    
-    var body: some ReducerOf<Self> {
-        BindingReducer()
-        
-        Reduce { state, action in
-            switch action {
-            case .binding, .delegate:
-                return .none
-            }
-        }
-    }
-}
-
-
-import SwiftUI
-
-struct SettingList: View {
-    @Bindable var store: StoreOf<SettingListFeature>
-    
-    var body: some View {
+    public var body: some View {
         List {
             Section {
                 Button {
@@ -170,6 +108,10 @@ struct SettingList: View {
             .tint(.black)
         }
         .navigationTitle("더보기")
+    }
+    
+    public init(store: StoreOf<SettingListFeature>) {
+        self.store = store
     }
 }
