+++ KuringPackage/Sources/UIKit/SettingsUI/WebView.swift
@@ -1,0 +1,25 @@
+import UIKit
+import WebKit
+import SwiftUI
+
+// TODO: Common UIKit
+public struct WebView: UIViewRepresentable {
+    public var urlToLoad: String
+    
+    public func makeUIView(context: Context) -> WKWebView {
+        guard let url = URL(string: self.urlToLoad) else {
+            return WKWebView()
+        }
+        let webView = WKWebView()
+        webView.load(URLRequest(url: url))
+        return webView
+    }
+    
+    public func updateUIView(_ uiView: WKWebView, context: UIViewRepresentableContext<WebView>) {
+        
+    }
+    
+    public init(urlToLoad: String) {
+        self.urlToLoad = urlToLoad
+    }
+}
